setTimeout instead of setInterval
I think that for our use of setInterval, it may be more correct to use setTimeOut. Because setTimeOut executes the function after a certain time, unlike setInterval that executes the function every so often. With the use of setTimeOut it is not necessary to eliminate the timer in each iteration, it would only be necessary to create a new timer.